+++ dotnet/App.axaml.cs
@@ -3,8 +3,8 @@ using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Data.Core;
 using Avalonia.Data.Core.Plugins;
 using Avalonia.Markup.Xaml;
-using dotnet.ViewModels;
-using dotnet.Views;
+using dotnet.Common.Presentation;
+using dotnet.Features.Dashboard.Presentation;
 
 namespace dotnet;
 
@@ -20,7 +20,7 @@ public partial class App : Application
             if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
                 desktop.MainWindow = new MainWindow();
             else if (ApplicationLifetime is ISingleViewApplicationLifetime singleView)
-                singleView.MainView = new MainSingleView();
+                singleView.MainView = new DashboardSingleView();
 
             base.OnFrameworkInitializationCompleted();
     }
