+++ dotnet/Features/Dashboard/Presentation/DashboardSingleView.axaml.cs
@@ -1,0 +1,13 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace dotnet.Features.Dashboard.Presentation;
+
+public partial class DashboardSingleView : UserControl
+{
+    public DashboardSingleView()
+    {
+        InitializeComponent();
+    }
+}
