--- dotnet/ViewModels/ViewModelBase.cs
@@ -1,7 +1,0 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-
-namespace dotnet.ViewModels;
-
-public class ViewModelBase : ObservableObject
-{
-}
