--- dotnet/Views/MainSingleView.axaml
@@ -1,11 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:rpiTestAppViews="clr-namespace:dotnet.Views"
-             mc:Ignorable="d"
-             d:DesignWidth="800"
-             d:DesignHeight="450"
-             x:Class="dotnet.Views.MainSingleView">
-    <rpiTestAppViews:MainView />
-</UserControl>
