--- dotnet/Views/MainSingleView.axaml.cs
@@ -1,13 +1,0 @@
-using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace dotnet.Views;
-
-public partial class MainSingleView : UserControl
-{
-    public MainSingleView()
-    {
-        InitializeComponent();
-    }
-}
