--- dotnet/Views/MainView.axaml.cs
@@ -1,15 +1,0 @@
-using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-using SkiaSharp;
-using Svg.Skia;
-
-namespace dotnet.Views;
-
-public partial class MainView : UserControl
-{
-    public MainView()
-    {
-        InitializeComponent();
-    }
-}
