--- dotnet/Views/MainWindow.axaml
@@ -1,12 +1,0 @@
-<Window xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:dotnetViews="clr-namespace:dotnet.Views"
-        mc:Ignorable="d"
-        d:DesignWidth="582"
-        d:DesignHeight="372"
-        x:Class="dotnet.Views.MainWindow"
-        Title="dotnet">
-    <dotnetViews:MainView />
-</Window>
