+++ dotnet/dotnet.csproj
@@ -9,7 +9,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Models\" />
     <AvaloniaResource Include="Assets\**" />
   </ItemGroup>
 
@@ -27,5 +26,6 @@
     <PackageReference Include="Serilog" Version="4.1.0" />
     <PackageReference Include="SkiaSharp.NativeAssets.Linux" Version="2.88.8" />
     <PackageReference Include="Svg.Skia" Version="2.0.0.1" />
+    <PackageReference Include="xunit" Version="2.9.2" />
   </ItemGroup>
 </Project>
